requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
iniconfig==1.1.1
packaging==21.3
paramiko==2.9.2
Pint==0.18
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyfrc==2022.0.2
PyNaCl==1.5.0
pynetconsole==2.0.2
pynetworktables==2021.0.0
pyntcore==2022.2.1.2
pyparsing==3.0.7
pytest==6.2.5
robotpy==2022.0.33
robotpy-ctre==2022.0.5
robotpy-hal==2022.2.1.1
robotpy-halsim-gui==2022.2.1.0
robotpy-installer==2022.0.2
robotpy-wpilib-utilities==2022.0.1
robotpy-wpimath==2022.2.1.4
robotpy-wpiutil==2022.2.1.1
six==1.16.0
toml==0.10.2
wpilib==2022.2.1.5
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.5.62
Werkzeug==2.0.2